requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
pyodbc
pytest
requests
passlib[bcrypt]
python-jose[cryptography]
python-multipart
xlwt
lxml
icecream